\name{fletcher2017data-package}

\alias{fletcher2017data-package}
\alias{fletcher2017data}
\alias{fletcher}
\alias{clustered}

\docType{package}

\title{
\packageTitle{fletcher2017data}
}

\description{
\packageDescription{fletcher2017data}
}

\details{

The dataset \code{fletcher} contains the raw, unfiltered data (28284 genes, 849 cells).

The dataset \code{clustered} contains the filtered data (1000 most variable genes, 747 cells), in which \code{zinbwave} dimensionality reduction and \code{RSEC} clustering has been performed.

}

\author{
Davide Risso
}

\references{
Fletcher, Das, Gadye, Street, Baudhuin, Wagner, Cole, Flores, Choi, Yosef, Purdom, Dudoit, Risso, Ngai. Deconstructing Olfactory Stem Cell Trajectories at Single-Cell Resolution. Cell Stem Cell, 20(6):817-830 (2017).
}